requirements.txt
argh==0.26.2
mido==1.2.9
numpy==1.20.2
pyFluidSynth==1.3.0
tqdm==4.60.0
requests=2.25.1